+++ HISTORY.rst
@@ -9,6 +9,8 @@ As of 1.0.0, `semantic versioning <http://semver.org/>`__ is used.
 +++++++++
 released 2013-XX-XX
 
+- add Mobileclient
+- remove broken Webclient.{create_playlist, change_playlist, copy_playlist, search, change_playlist_name}
 - add support for streaming All Access songs
 - add Webclient.get_registered_devices
 - add a toggle to turn off validation per client